FEATURE - start RunDB
Rationale:  add a Run table to the local sqlite db on the tof cpu.
The reason is the following - the tof cpu "knows" best which run id
to assigned, based on its local data.
To avoid confusion, we will start automaticallyl assigned run ids with 1000.
The usage of this db is still work-in-progress.