FIX - no global trigger setting in combo mode.
RATIONALE: This was a misunderstanding. The "global" means that
the trigger denoted as "global" is not trace suppressed, but reads
out all paddles. It has nothing to do with the combo mode, basically
all the triggers can be combined independently, maybe not the
configurable ones.

FIX - fixes tiu_busy_ignore register readout for the use with the
      new register