FIX - addresses a few issues with the event builder
The issues are related to the refactoring of TofEvent and the
late addition of some "interesting event" selection

* Events have been sent twice. This was a simple bug and has been
  fixed.

ISSUE: It is still a bit unclear, what is the best approach to the
question of which events to send, which events to write to disk. The
angle important here is where do we make that decision. event_builder
or global_data_sink.